match_color8: Signedness fix.

+++ src/exp-txt.c
@@ -21,7 +21,7 @@
  *  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
-/* $Id: exp-txt.c,v 1.15 2006/02/10 06:25:37 mschimek Exp $ */
+/* $Id: exp-txt.c,v 1.16 2006/05/22 09:07:53 mschimek Exp $ */
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -261,9 +261,9 @@ match_color8(vbi_rgba color)
 	int i, d, imin = 0, dmin = INT_MAX;
 
 	for (i = 0; i < 8; i++) {
-		d  = ABS(       (i & 1) * 0xFF - VBI_R(color));
-		d += ABS(((i >> 1) & 1) * 0xFF - VBI_G(color));
-		d += ABS( (i >> 2)      * 0xFF - VBI_B(color));
+		d  = ABS((int)(       (i & 1) * 0xFF - VBI_R(color)));
+		d += ABS((int)(((i >> 1) & 1) * 0xFF - VBI_G(color)));
+		d += ABS((int)( (i >> 2)      * 0xFF - VBI_B(color)));
 
 		if (d < dmin) {
 			dmin = d;
